requirements.txt
fastapi
uvicorn[standard]
streamlit
pdfplumber
sentence-transformers
faiss-cpu
numpy
pandas
rank_bm25
groq
python-dotenv
requests
python-multipart